feat: logged_in state
